revert restructure
